\name{superpc.news}

\alias{superpc.news}

\title{Display the \pkg{superpc} Package News}

\description{
    Function to display the log file \code{NEWS} of updates of the \pkg{superpc} package.
}

\usage{
    superpc.news(...)
}

\arguments{
    \item{\dots}{Further arguments passed to or from other methods.}
}

\value{
    None.
}

\references{
    \itemize{
        \item E. Bair and R. Tibshirani (2004).
              "\emph{Semi-supervised methods to predict patient survival from gene expression data}."
              PLoS Biol, 2(4):e108.
        \item E. Bair, T. Hastie, D. Paul, and R. Tibshirani (2006).
              "\emph{Prediction by supervised principal components}."
              J. Am. Stat. Assoc., 101(473):119-137.
    }
}

\author{
    \itemize{
        \item "Eric Bair, Ph.D."
        \item "Jean-Eudes Dazard, Ph.D."
        \item "Rob Tibshirani, Ph.D."
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D."
}

\note{
    End-user function. 
}

\keyword{documentation}
